Proyecto terminado

+++ src/main/java/services/ItemsServices.java
@@ -29,39 +29,35 @@ public class ItemsServices {
     public List<Item> getAll(){
         return dao.getDAOItems().getAll();
     }
-    public Item get(int id){
-        return dao.getDAOItems().get(id);
-    }
+
 
     public boolean save(Item item){
         return dao.getDAOItems().save(item);
 
     }
-    public int deletePurchasesAndItem(Item it){
-        return dao.getDAOItems().deletePurchasesAndItem(it.getIdItem());
-    }
 
-    public int deleteItem(int id){
-        return dao.getDAOItems().deleteItem(id);
-    }
-
-    public int checkItemReviews(int id){
-        return dao.getDAOItems().checkItemReview(id);
-    }
     public int checkDelete(int id,boolean confirm){
-        int confirmacion = -5;
+        int confirmacion = -6;
         int resultado = dao.getDAOItems().checkItemReview(id);
         int resultadoPurchases = dao.getDAOPurchases().getPurchasesByItemId(id);
         if (resultado > 0){
             confirmacion = -1;
         }else if(resultadoPurchases > 0 && !confirm){
-            confirmacion = -2;
-        }else if (resultadoPurchases > 0 && confirm){
             confirmacion = -3;
-            dao.getDAOItems().deletePurchasesAndItem(id);
+        }else if (resultadoPurchases > 0 && confirm){
+            if (dao.getDAOItems().deletePurchasesAndItem(id) == -2){
+                confirmacion = dao.getDAOItems().deletePurchasesAndItem(id);
+            }else{
+                confirmacion = -4;
+            }
+
         } else if(resultadoPurchases == 0){
-            confirmacion = -4;
-            dao.getDAOItems().deleteItem(id);
+            if ( dao.getDAOItems().deleteItem(id) == -2){
+                confirmacion =  dao.getDAOItems().deleteItem(id);
+            }else{
+                confirmacion = -5;
+            }
+
         }
